#4[SPIKE] Add BetDAO implementation

+++ src/main/java/com/epam/model/Bet.java
@@ -9,6 +9,6 @@ public class Bet {
     private long id;
     private long offerId;
     private long personId;
-    private int amount;
-    private int gain;
+    private long amount;
+    private float gain;
 }
